TestRun: Added tests/tc/test_tc0.ref as sanity test and uncovered bugs:
1. Unable to use the type 'int'. Cub's syntax error barfs (had to replace int -> inty)
2. Syntax error checking itself is painful. Have to look at ldump.txt to guess what the syntax error is.
3. (2) needs to be properly fixed for cub
4. multi line comments don't seem to work
5. constants like 1 and 0 when assigned to a var don't seem to work
6. array declarations don't seem to work
7. simple 'in' context does not seem to work

To get around these problems I created a tests/tc/test_tc0 that passes
with a couple of workarounds I made. The ldump.txt and ydump.txt (along with out)
should now be very instructive in debugging what's going on ...

Modified the run script to pass arguements (tests) and added a last_run
that I still need to figure out how to automate.