Showing that expression-evual does not include l-values yet
Sequence expression is a null inside parenthesis, or a proper sequence
(see tig.yac) but it is missing the concept of an l-value (see the tiger
reference lang constructs in the appendix).

This is highlighted by pointing run to tests/tc/test_tc0.ref
where the 'in' clause simply calls an expression without param.
(Another test would be for an 'in' clause to simply refer to an
l-value as well).

This will need to be fixed next for test_tc0.ref to pass.