Add additional tests (queens.mod2.tig and .ref)
These tests show changes I need to make to get the abstract syntax
tree to move forward and shows things that still need to be worked on.

Examples : For kwd preceded with LParan, Functional calls with string literals,
re-examination of SEND_FALSE_FUNCTION, general yyerror(x) not able to print
out the reason for the error etc. These tests also expose to-be-tested features
like : mutually recursive function declarations (and type declarations) which should
all be a single decl of that kind, nested for loops etc. All of these are also
great examples of multiple types of scope (addressed in later stages)

Still focus on getting queens.mod.tig to work (and still understand the
reduce-reduce conflict with array decl). Basically diffs from the main
queens.tig show progressive 'fixes' all of which should be addressed.