Fix the lparan-for problem
Handle for in its own construct. The realization is that
a for kwd in cub doesn't concern itself with what precedes
it ($, [], ( etc) but that it has a [] following itself,
to distinguish it from ident. Having said that, I still show
3 different forms of for which could be replaced with 1 construct.

Also realization that a certain series of tests should be based
on all cases of a constuct usage (like a set of for-loop tests)
These could be part of a simple compilable program and should
show passing as well as error cases (with the compiler reacting
correctly to both scenarios)