Add pure_string definition and add expseq(explist) to point to expv (expliterals)
This solves the if definition in queen.mod.tig in the first function's definition.
It exposes that func call params don't recognize the same if defn because
parameter_list only refers to expv, but only expseq contains explist->ifdef_decl

It would be erroneous to have parameter_list call expseq because explist only
evaluates a semi-col but parameter_list needs expressions with comma. Also expseq
is defined correctly to be lparan (valued-expr(s) ) rparan.

The concept of expressions therefore needs a re-write in tig.yac. Also a cub program is
an 'expression' as defined in the Appendix of tigerbook.pdf. Thus type% 'expression'
has to be introduced/rewritten. (expuv also needs to be a formal %type)

Fixed some minor print typos